User Management update

+++ Project/UserManagement.cs
@@ -17,6 +17,7 @@ namespace Project
             this.StartPosition = FormStartPosition.CenterScreen;
         }
 
+        //FIELDS
         StudentUserRepository repository = new StudentUserRepository();
         private DataGridViewRow selectedRow;
         private int selectedStudentId = -1;
@@ -35,26 +36,41 @@ namespace Project
             user.FullName = selectedRow.Cells["FullName"].Value?.ToString() ?? "";
             user.Email = selectedRow.Cells["Email"].Value?.ToString() ?? "";
             user.Password = selectedRow.Cells["Password"].Value?.ToString() ?? "";
+
             repository.UpdateStudent(user);
             MessageBox.Show("User updated successfully!");
             LoadUsers();
         }
 
+        //DELETE
+        private void ButtonDelete_Click(object sender, EventArgs e) {
+            if (selectedStudentId == -1) return;
+            StudentUser user = repository.Get(selectedStudentId);
+            bool isDeleted = repository.DeleteStudent(user);
+            if (isDeleted) MessageBox.Show("Successfully Removed");
+            LoadUsers();
+        }
+
+        //GET INDEX
         private void DataGridViewUserList_CellClick_1(object sender, DataGridViewCellEventArgs e) {
             if (e.RowIndex < 0) return;
             selectedRow = DataGridViewUserList.Rows[e.RowIndex];
             selectedStudentId = Convert.ToInt32(selectedRow.Cells["StudentId"].Value);
         }
 
+        //LOAD USERS
         private void LoadUsers() {
             List<StudentUser> users = repository.GetAll();
             DataGridViewUserList.DataSource = users;
         }
 
-        private void ButtonDelete_Click(object sender, EventArgs e) {
-            StudentUser user = repository.Get(selectedStudentId);
-            bool isDeleted = repository.DeleteStudent(user);
-            if (isDeleted) MessageBox.Show("Successfully Removed");
+        private void ButtonCreate_Click(object sender, EventArgs e) {
+            StudentUser user = new StudentUser();
+            user.FullName = string.Empty;
+            user.Email = string.Empty;
+            user.Password = string.Empty;
+
+            repository.Add(user);
             LoadUsers();
         }
     }
